fix(client-app/redux/*) folder name change, reducer refactoring
